[REFACT] 수정사항 반영
