+++ app/src/main/java/com/umc/edison/data/model/bubble/PositionBubbleEntity.kt
@@ -1,7 +1,7 @@
 package com.umc.edison.data.model.bubble
 
 data class PositionBubbleEntity(
-    val id : String,
+    val id: String,
     val x: Float,
     val y: Float,
     val group: Int
