+++ app/src/main/java/com/umc/edison/presentation/model/KeywordBubbleModel.kt
@@ -2,18 +2,18 @@ package com.umc.edison.presentation.model
 
 import com.umc.edison.domain.model.bubble.KeywordBubble
 
-data class KeywordBubbleModel (
-    val bubble : BubbleModel,
-    val similarity : Float
+data class KeywordBubbleModel(
+    val bubble: BubbleModel,
+    val similarity: Float
 )
 
-fun KeywordBubble.toPresentation():KeywordBubbleModel{
+fun KeywordBubble.toPresentation(): KeywordBubbleModel {
     return KeywordBubbleModel(
-        bubble=  bubble.toPresentation(),
+        bubble = bubble.toPresentation(),
         similarity = similarity
     )
 }
 
-fun List<KeywordBubble>.toPresentation():List<KeywordBubbleModel>{
-    return map{ it.toPresentation() }
+fun List<KeywordBubble>.toPresentation(): List<KeywordBubbleModel> {
+    return map { it.toPresentation() }
 }
