+++ app/src/main/java/com/umc/edison/ui/components/SearchBar.kt
@@ -93,72 +93,3 @@ fun SearchBar(
         }
     )
 }
-
-@Composable
-fun KeywordMapBar(
-    value: String,
-    onValueChange: (String) -> Unit,
-    onMap: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier.fillMaxWidth(),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(48.dp)
-                .clip(RoundedCornerShape(100.dp))
-                .background(Gray300)
-                .padding(horizontal = 12.dp, vertical = 8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            BasicTextField(
-                value = value,
-                onValueChange = onValueChange,
-                keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Search),
-                keyboardActions = KeyboardActions(onSearch = { onMap() }),
-                textStyle = MaterialTheme.typography.bodySmall.copy(color = Gray800),
-                modifier = Modifier.weight(1f),
-                singleLine = true,
-                decorationBox = { innerTextField ->
-                    Box(
-                        contentAlignment = Alignment.CenterStart
-                    ) {
-                        if (value.isEmpty()) {
-                            Text(
-                                text = "키워드를 입력하세요",
-                                style = MaterialTheme.typography.bodySmall,
-                                color = Gray600
-                            )
-                        }
-                        innerTextField()
-                    }
-                }
-            )
-        }
-
-        Spacer(modifier = Modifier.height(30.dp))
-
-        Box(
-            modifier = Modifier
-                .clip(RoundedCornerShape(100.dp))
-                .background(Color.White)
-                .border(
-                    2.dp,
-                    Brush.horizontalGradient(listOf(Color(0xFFFFFBE4), Color(0xFFFF6FDF), Color(0xFF85C9FF))),
-                    RoundedCornerShape(100.dp)
-                )
-                .clickable { onMap() },
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                text = "키워드로 매핑",
-                style = MaterialTheme.typography.titleSmall,
-                color = Gray800,
-                modifier = Modifier.padding(horizontal = 20.dp, vertical = 15.dp)
-            )
-        }
-    }
-}
